requirements.txt
python-telegram-bot==21.6
requests
uvicorn
fastapi
dotenv
openai
thefuzz[speedup]
pandas